requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cloudinary==1.44.1
cryptography==45.0.6
dj-database-url==3.0.1
Django==5.2.5
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
ecdsa==0.19.1
firebase_admin==7.1.0
frozenlist==1.7.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.3.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
MarkupSafe==3.0.2
msgpack==1.1.1
multidict==6.6.4
pillow==11.3.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.1
python-http-client==3.3.7
reportlab==4.4.3
requests==2.32.4
rsa==4.9.1
sendgrid==6.12.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
twilio==9.7.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.20.1
